Addressed inconsistency & deadlock in topologymanager.
The usageCount was not consistently incremented/decremented based on OSGi event.
In addition the usageCount check was missing in the loop where senders/receivers are created,
resulting in resources being created when the count was 0. This could result in a deadlock
when in parallel the need for these resources would re-appear.